Move batch saving to LocalEnsemble
