+++ src/ert/storage/local_ensemble.py
@@ -9,7 +9,7 @@ from datetime import datetime
 from functools import cache, cached_property, lru_cache
 from multiprocessing.pool import ThreadPool
 from pathlib import Path
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, ClassVar
 from uuid import UUID
 
 import numpy as np
@@ -40,6 +40,21 @@ class EverestRealizationInfo(TypedDict):
     perturbation: int  # -1 means it stems from unperturbed controls
 
 
+class BatchDataframes(TypedDict, total=False):
+    realization_controls: pl.DataFrame | None
+    batch_objectives: pl.DataFrame | None
+    realization_objectives: pl.DataFrame | None
+    batch_constraints: pl.DataFrame | None
+    realization_constraints: pl.DataFrame | None
+    batch_bound_constraint_violations: pl.DataFrame | None
+    batch_input_constraint_violations: pl.DataFrame | None
+    batch_output_constraint_violations: pl.DataFrame | None
+    batch_objective_gradient: pl.DataFrame | None
+    perturbation_objectives: pl.DataFrame | None
+    batch_constraint_gradient: pl.DataFrame | None
+    perturbation_constraints: pl.DataFrame | None
+
+
 class _Index(BaseModel):
     id: UUID
     experiment_id: UUID
@@ -68,6 +83,21 @@ class LocalEnsemble(BaseMode):
     Manages multiple realizations of experiments, including different sets of
     parameters and responses.
     """
+
+    BATCH_DATAFRAMES: ClassVar[list[str]] = [
+        "realization_controls",
+        "batch_objectives",
+        "realization_objectives",
+        "batch_constraints",
+        "realization_constraints",
+        "batch_bound_constraint_violations",
+        "batch_input_constraint_violations",
+        "batch_output_constraint_violations",
+        "batch_objective_gradient",
+        "perturbation_objectives",
+        "batch_constraint_gradient",
+        "perturbation_constraints",
+    ]
 
     def __init__(
         self,
@@ -1247,6 +1277,12 @@ class LocalEnsemble(BaseMode):
         ]
 
         return params_and_responses[column_order]
+
+    def save_dataframes(cls, dataframes: BatchDataframes, ensemble_path: Path) -> None:
+        for df_name in cls.BATCH_DATAFRAMES:
+            df = dataframes.get(df_name)
+            if isinstance(df, pl.DataFrame):
+                df.write_parquet(ensemble_path / f"{df_name}.parquet")
 
 
 async def _read_parameters(
